requirements.txt
# Question Generator Dependencies
langgraph>=0.2.0
langchain>=0.3.0
langchain-openai>=0.2.0
matplotlib>=3.7.0
numpy>=1.24.0
pyyaml>=6.0
python-dotenv>=1.0.0
